Change from info logging to printing to the console in Execute class of files.py.
Logging to info here causes a duplicated logging header to be appended to output that was already formatted for logging.  Switching to printing preserves the formatting of existing log output while still allowing standard output from binaries and PowerShell (such as Write-Verbose) to still be written to the console.

Also changed formatter from digit to string as breaking out of Glazier was returning 'None'.

PiperOrigin-RevId: 273586473