Extend SetStage to allow terminal stages.
Normally, the 'active' stage is exited when another call to StartStage happens. This will never happen for the final build stage (the "terminal" stage), since it is by definition the last stage.

Extend the action with a bool field to indicate the last expected stage, which will start and then immediately end the stage. This allows clients to detect that all stages are complete. (We could also create an EndStage action, but this seems unnecessary overhead when one call can do everything.)

PiperOrigin-RevId: 319831413