cmd.Wait() for a non 0 exit code is filled with metadata (it's assuming any non 0 exit code is bad), by pinning verify on this being nil means we are doing the same
```
The returned error is nil if the command runs, has no problems copying stdin, stdout, and stderr, and exits with a zero exit status.
```

Do some additional parsing of res.ExitErr first to not bail on verification if ExitErr contains a "bad" exit status and let the additional checks determine if the run was bad.

PiperOrigin-RevId: 463916936